fix: move search input into validators tabs
